feat: 마이페이지 리뷰 반영
- CartOptionModal.tsx 명칭 오류 수정
- MypageSideNav.tsx router.refresh() 삭제
- 공통 타입 정리
