+++ apps/client/src/features/product/components/product-detail/ProductDetailSummary.tsx
@@ -2,7 +2,7 @@
 
 import { fetchAddCart } from "@/features/cart/api";
 import { CartItem } from "@/features/cart/model";
-import { usePostWishlist } from "@/features/mypage/hook/useWishlist";
+import { deleteWishListAPI, postWishListAPI } from "@/features/mypage/api";
 import { ProductOption } from "@/features/product/model";
 import {
   findProductDetailId,
@@ -23,6 +23,7 @@ type Props = {
   sellingPrice: number;
   discountRate: number;
   mainImages: string;
+  isWished: boolean;
   options: ProductOption[];
 };
 
@@ -35,6 +36,7 @@ export default function ProductDetailSummary({
   discountRate,
   mainImages,
   options,
+  isWished,
 }: Props) {
   // 옵션 선택
   const [selectedColor, setSelectedColor] = useState<string>("");
@@ -95,18 +97,24 @@ export default function ProductDetailSummary({
   });
 
   // 찜하기
-  const { addWishlist, isLoading, error } = usePostWishlist();
+  // 추후 api 수정 후 변경 예정/ 리뷰 달지 말아주세요~!
+  const [isWish, setIsWish] = useState(isWished);
   const handleWishClick = async () => {
     try {
-      await addWishlist(productId);
-      alert(`${productId}번이 찜~`);
-    } catch {
-      console.error(error);
+      if (isWish) {
+        // console.log("찜하기에서 id는", productId);
+        await deleteWishListAPI([productId]);
+        setIsWish(false);
+      } else {
+        await postWishListAPI(productId);
+        setIsWish(true);
+      }
+    } catch (err) {
+      console.error(err);
     }
   };
 
   // 장바구니 담기
-  // 추후 찜 여부 구분값 생기면 수정 예정(리뷰달지마!)
   const handleCartClick = async () => {
     if (!selectedColor || !selectedSize) {
       setModalMessage("옵션을 선택해주세요.");
@@ -300,11 +308,11 @@ export default function ProductDetailSummary({
 
       <article className="flex items-center gap-2.5">
         <div
-          className="group bg-text-disabled flex h-8 w-8 cursor-pointer items-center justify-center rounded-full hover:bg-neutral-900 lg:mx-3.5"
+          className={`${isWish ? "bg-neutral-900" : "bg-text-disabled"} group flex h-8 w-8 cursor-pointer items-center justify-center rounded-full hover:bg-neutral-900 lg:mx-3.5`}
           onClick={handleWishClick}
         >
           <svg
-            className="fill-white group-hover:fill-[#FEC300]"
+            className={`${isWish ? "fill-[#FEC300]" : "fill-white"} group-hover:fill-[#FEC300]`}
             width="18"
             height="17"
             viewBox="0 0 18 17"
