+++ apps/client/src/shared/type/model.ts
@@ -1,3 +1,4 @@
+// page response type
 export type PageResponse<T> = {
   content: T[];
   pageNumber: number;
@@ -5,4 +6,15 @@ export type PageResponse<T> = {
   totalElements: number;
   totalPages: number;
   last: boolean;
+};
+
+// product options type
+export type ProductOptionSize = {
+  size: string;
+  soldOut: boolean;
+  productDetailId: number;
+};
+export type ProductOption = {
+  color: string;
+  sizes: ProductOptionSize[];
 };
